Refactor of Conversation Model and adding query to find conversation
TBD: creating new conversation in case if users don't have actually conversation

+++ backend/src/main/java/com/umbra/umbralink/dto/ConversationDto.java
@@ -9,7 +9,7 @@ public class ConversationDto {
     private String otherUser;
     private Long otherUserId;
     private String lastMessage;
-    private int conversationId;
+    private Long conversationId;
     private MessageState state;
     private Boolean isLastMessageSender;
     private UserStatus status;
