+++ backend/src/main/java/com/umbra/umbralink/dto/conversationData/ConversationDataDto.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class ConversationDataDto {
     private List<ConversationMessageDto> messages;
-    private int conversationId;
+    private Long conversationId;
     private Long receiverId;
     private String receiverName;
 }
